feat(homepage): merge Solutions into home page; /solutions redirects to /
- Home page now has hero, stats bar, 3 use-case sections with visuals,
  open standards pillars, How it works, popular/featured app showcase,
  CTA banner, and footer
- App showcase placed between How it works and CTA banner
- /solutions redirects to / (backward-compatible link)
